many: make unit tests pass with nosecboot tags
Some distros may want to run the unit tests, but the secboot package is not
avaialble there. Make sure that the unit tests are passing when nosecboot tag is
passed.
